+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [DGPB-7161] Modify Gosec function deleteResourceIfExistsDcos to delete collections policies
 * [PQA-146] New steps to download and upload k8s universe
 * [DGPB-8233] Modify function to copy directories in/from pod
+* [DGPB-8260] Add support to Governance requests in k8s
 
 ## 0.20.0 (December 17, 2020)